Refactor the generic parts of the client into an abstract class
