+++ RELEASE_NOTES.md
@@ -1,9 +1,7 @@
 
 ### New Features
 
- - Add the ability to specify attributes, addresses, form_fields, and authentication when creating customers (#67)
+
 
 ### Fixes
 
- - Add missing Customer properties (`notes`, `tax_exempt_category`, `accepts_product_review_abandoned_cart_emails`, 
-   `store_credit_amounts`, `origin_channel_id`, and `channel_ids`)
